#169 Fix of test and renaming of files
