+++ src/main.ts
@@ -8,7 +8,7 @@ export * from './navigation-drawer/navigation-drawer.component';
 export * from './icon/icon.component';
 export * from './button/button.directive';
 export * from './list/list.component';
-export * from './tabbar/tab.component';
+export * from './tabbar/tabbar.component';
 export * from './carousel/carousel.component';
 export * from './input/input.directive';
 export * from './checkbox/checkbox.component';
